requirements.txt
crawl4ai>=0.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
asyncio
aiohttp>=3.8.0
playwright>=1.40.0 
solvecaptcha-python>=1.0.0 